+++ core/lib/circuit/src/witness/deposit.rs
@@ -248,7 +248,7 @@ impl DepositWitness<Bn256> {
                 pub_nonce: Some(Fr::zero()),
                 new_pub_key_hash: Some(Fr::zero()),
                 valid_from: Some(Fr::zero()),
-                valid_until: Some(Fr::from_str(&u64::MAX.to_string()).unwrap()),
+                valid_until: Some(Fr::from_str(&u32::MAX.to_string()).unwrap()),
             },
             before_root: Some(before_root),
             after_root: Some(after_root),
